Added MageEnemy, DraynorFishnCook. Fixed VialCrafter.
Added MageEnemy
first Param is enemy Id, second param is offset for distance to travel otherwise it'll manually assume 0 and always try to walk back to original starting position when it is pulled off. Must manually edit the script to change the spell, Currently set to "Fire bolt"

DraynorFishnCook is a low level (Level 1+) fishing script that requires an axe, net, and tinderbox that fishes, chops trees, lights a fire, and cooks the catch on the fire in Draynor. Used to get low level fishers into the mid levels for Catherby scripts or higher.

Fixed Vialcrafter. Removed the extra drops from the soda ash and molten glass, added an extra bucket pickup, and doubled the wait time after dropping edible seaweed. Tests so far are not conclusive but very promising for a potential fix.